Set reservation to zero

+++ tool/state-processor/src/type_registry.rs
@@ -224,7 +224,7 @@ impl Encode for Data {
 			Data::Raw(ref x) => {
 				let l = x.len().min(32);
 				let mut r = vec![l as u8 + 1; l + 1];
-				r[1..].copy_from_slice(&x[..l as usize]);
+				r[1..].copy_from_slice(&x[..l]);
 				r
 			},
 			Data::BlakeTwo256(ref h) => once(34u8).chain(h.iter().cloned()).collect(),
@@ -309,7 +309,7 @@ impl Encode for IdentityFields {
 impl Decode for IdentityFields {
 	fn decode<I: Input>(input: &mut I) -> Result<Self, Error> {
 		let field = u64::decode(input)?;
-		Ok(Self(<BitFlags<IdentityField>>::from_bits(field as u64).map_err(|_| "invalid value")?))
+		Ok(Self(<BitFlags<IdentityField>>::from_bits(field).map_err(|_| "invalid value")?))
 	}
 }
 #[bitflags]
